Major Additions: Added login and Rating controllers . Major Additions: Added Authentiation Filter and Authentication Util to create and retrive JWT tokens.
Bugs : fixed bugs in user and show controller and their corresponding
daos and services.
Fixed the Update function in AbstractDAO to use session.update than
using SaveOrUpdate as it was leading to first trying to save then update
which was leading to uniques constraint violation.

Fixed Dao Impls of user, show and rating which were earlier retriving
data using Criteria. They are now Using HQL , For some reason Criteria
API was returing more than one result even if there is only one matching
record accoring to the critera and even if there is only a single row
present in the databse. This problem was happening specifically when
retriving related collections withing entities for example , ShowDetails
has a list of related Directors , So While retriving even if there is
only one row in Direectors table the returned result was having a list
of same directors repeated several times . Couldn't find the issue so
switched to using HQL whcih solved the issue.


Refactoring  : Refactored the Details dto to be compatible to Jackson ,
as earlier it was created to be compatible with Gson library.
Refactoring : Moved Create User function from User Controller to Login
Controller