Added capablity to get top rated titles based on Imdb Ratings or ShowFlix Ratings
Added showFlix Rating member to Showdetails and Details Beans.
USAGE of showFlix: Whenever a showFlix user rates a title a
corresponding entry is created in Rating table and simoultaneously the
average rating of that title is updated in the ShowDetails table where
the updated rating is now stored in showflixRating instance member
field.